move functions, use config option to enable mvc, do so before importing driver
